requirements.txt
absl-py==0.11.0
alabaster==0.7.12
arch==4.15
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==20.2.0
Babel==2.8.0
backcall==0.2.0
bayesian-changepoint-detection @ file:///home/cfreeman/Documents/NextIT/jair_work/bayesian_changepoint_detection
bleach==1.5.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cmdstanpy==0.9.5
convertdate==2.2.2
creme==0.6.1
cycler==0.10.0
Cython==0.29.17
decorator==4.4.2
defusedxml==0.6.0
docutils==0.16
Donut @ file:///home/cfreeman/Documents/NextIT/jair_work/donut
entrypoints==0.3
ephem==3.7.7.1
fbprophet==0.7.1
filelock==3.0.12
frozendict==1.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
holidays==0.10.3
html5lib==0.9999999
idna==2.10
idx2numpy==1.2.3
imagesize==1.2.0
importlib-metadata==2.0.0
iniconfig==1.1.1
ipykernel==5.3.4
ipyparallel==6.3.0
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
korean-lunar-calendar==0.2.1
LunarCalendar==0.0.9
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.2
matrixprofile-ts==0.0.9
mistune==0.8.4
mmh3==2.5.1
natsort==7.0.1
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.2
nose==1.3.7
notebook==6.1.4
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.3
parso==0.7.1
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pkg-resources==0.0.0
pluggy==0.13.1
pmdarima==1.7.1
prometheus-client==0.8.0
prompt-toolkit==3.0.8
property-cached==1.6.4
protobuf==3.13.0
ptyprocess==0.6.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.2
PyMeeus==0.3.7
pyparsing==2.4.7
pyrsistent==0.17.3
pystan==2.19.1.1
pytest==6.1.1
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
rpy2==3.3.6
rsa==4.6
scikit-learn==0.23.2
scikit-posthocs==0.6.6
scipy==1.5.3
seaborn==0.11.0
semver==2.13.0
Send2Trash==1.5.0
setuptools-git==1.2
six==1.15.0
sklearn==0.0
snowballstemmer==2.0.0
sortedcontainers==2.2.2
Sphinx==3.2.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
statsmodels==0.11.1
tensorboard==1.6.0
tensorboard-plugin-wit==1.7.0
tensorflow==1.6.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
TFSnippet @ git+https://github.com/haowen-xu/tfsnippet.git@ecc0b4d1e610cf8cfa8c236857a7dabee27d5543
threadpoolctl==2.1.0
toml==0.10.1
tornado==6.0.4
tqdm==4.51.0
traitlets==4.3.3
tzlocal==2.1
UNKNOWN @ file:///home/cfreeman/Documents/NextIT/jair_work/pbad/src/utils/cython_utils
urllib3==1.25.11
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zhusuan @ git+https://github.com/thu-ml/zhusuan.git@4386b2a12ae4f4ed8e694e504e51d7dcdfd6f22a
zipp==3.4.0
